feat: Separate Lama listener into a separate class
- Separate Lama listener from the main class to improve maintainability and code organization
- Decouple packet listener and Bukkit event listener responsibilities
