+++ src/main/kotlin/de/danielmaile/mpp/world/aether/ListenerAether.kt
@@ -18,16 +18,11 @@
 package de.danielmaile.mpp.world.aether
 
 import de.danielmaile.mpp.aetherWorld
-import de.danielmaile.mpp.inst
-import de.danielmaile.mpp.packet.PacketEvent
-import de.danielmaile.mpp.packet.PacketListener
 import io.papermc.paper.event.entity.EntityMoveEvent
-import net.minecraft.network.protocol.game.ServerboundPlayerInputPacket
 import org.bukkit.Bukkit
 import org.bukkit.Location
 import org.bukkit.entity.Entity
 import org.bukkit.entity.LivingEntity
-import org.bukkit.entity.Llama
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
@@ -36,44 +31,8 @@ import org.bukkit.event.player.PlayerMoveEvent
 import org.bukkit.event.world.EntitiesLoadEvent
 import org.bukkit.potion.PotionEffect
 import org.bukkit.potion.PotionEffectType
-import org.bukkit.util.Vector
 
 class ListenerAether : Listener {
-
-    // TODO: Not registered
-    @PacketListener
-    fun steerLlama(event: PacketEvent<ServerboundPlayerInputPacket>) {
-        val player = event.player
-        val llama = player.vehicle ?: return
-        if (llama !is Llama) return
-        if (llama.location.world != aetherWorld()) return
-        if (llama.inventory.decor == null) return
-
-        val jump = event.packet.isJumping
-        if (llama.isOnGround && jump) {
-            llama.velocity = llama.velocity.clone().add(Vector(0.0, inst().configManager.llamaJumpHeight, 0.0))
-        }
-
-        val leftRight = event.packet.xxa
-        val forwardBackward = event.packet.zza
-
-        val horizontal = Vector(-forwardBackward.toDouble(), 0.0, leftRight.toDouble())
-        if (horizontal.length() > 0) {
-            // turn to face of vector in direction which the player is facing
-            horizontal.rotateAroundY(Math.toRadians((-player.location.yaw + 90f).toDouble()))
-
-            // scale vector
-            horizontal.normalize().multiply(inst().configManager.llamaSpeed)
-        }
-
-        val vertical = llama.getVelocity()
-        vertical.setX(0)
-        vertical.setZ(0)
-        llama.setVelocity(horizontal.add(vertical))
-
-        // rotation
-        llama.setRotation(player.eyeLocation.yaw, player.eyeLocation.pitch)
-    }
 
     /**
      * Adds slow falling effect to all mobs that are loaded in the aether.
